fix(fwd): clear the MSVC /W4 warnings in the forward library
With mne clean the Windows leg moved on to src/libraries/fwd and the tools built
on it, reporting 37 warnings.

Unlike the previous round these are not hiding bugs: the shadowing sites all
disambiguate explicitly already (this->nsol = nsol, surfs[s]->ntri), so the
locals were only obscuring the members by name. Renamed to nTriangles /
nSolutions / nPoints / bemMethod and, in the Berg-Scherg fit, nTerms / nFit /
lambdaFit / muFit. Member accesses through -> and this-> were left untouched and
verified afterwards: 61 remain, none rewritten.

The rest: float literals for the C4305 truncations, explicit casts for the
size_t narrowings, and FWD_BEM_IP_APPROACH_LIMIT is now a float since the member
it initialises is one.

Two of the C4996 reports were worth acting on rather than silencing.
mne_fiff_exp_set used localtime(), which returns a pointer into a shared static
buffer, so a concurrent caller could overwrite the result before strftime read
it; it now uses the reentrant variant. mne_list_bem moves off FILE* to QFile,
keeping printf-style formatting because the fixed-width layout is consumed by
other tools. The two sscanf("%d") calls in mne_show_fiff parse into an int with
no buffer involved, so they were only a deprecation; they now use QString::toInt
and actually report malformed input the same way.

Verified: clang build 0 warnings, 15/15 forward, BEM and sphere-model tests,
181/181 ctest overall.

+++ src/libraries/fwd/fwd_eeg_sphere_model_set.h
@@ -145,7 +145,7 @@ public:
     /** Number of models in this set. */
     int nmodel() const
     {
-        return models.size();
+        return static_cast<int>(models.size());
     }
 
 public:
